# openapi.yaml
openapi: 3.0.3
info:
  title: RAG Service
  version: 2.1.0
paths:
...
  /memory/push:
    post:
      tags:
      - memory
      summary: Push Memory
      operationId: push_memory_memory_push_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryPushRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryPushResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /memory/upload:
    post:
      tags:
      - memory
      summary: Upload Memory
      operationId: upload_memory_memory_upload_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_memory_memory_upload_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryUploadResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /memory/sessions:
    get:
      tags:
      - memory
      summary: List Memory Sessions
      operationId: list_memory_sessions_memory_sessions_get
      parameters:
      - name: app_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: App Id
      - name: wallet_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallet Id
      - name: data_wallet_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Data Wallet Id
      - name: session_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemorySessionList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - memory
      summary: Delete Memory Session
      operationId: delete_memory_session_memory_sessions_delete
      parameters:
      - name: app_id
        in: query
        required: true
        schema:
          type: string
          title: App Id
      - name: session_id
        in: query
        required: true
        schema:
          type: string
          title: Session Id
      - name: wallet_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallet Id
      - name: data_wallet_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Data Wallet Id
      - name: delete_files
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Delete Files
      - name: delete_vectors
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Delete Vectors
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryDeleteResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /memory/sessions/{memory_key}:
    delete:
      tags:
      - memory
      summary: Delete Memory Session By Key
      operationId: delete_memory_session_by_key_memory_sessions__memory_key__delete
      parameters:
      - name: memory_key
        in: path
        required: true
        schema:
          type: string
          title: Memory Key
      - name: wallet_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallet Id
      - name: data_wallet_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Data Wallet Id
      - name: delete_files
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Delete Files
      - name: delete_vectors
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Delete Vectors
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryDeleteResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /memory/{memory_key}/contexts:
    get:
      tags:
      - memory
      summary: List Memory Contexts
      operationId: list_memory_contexts_memory__memory_key__contexts_get
      parameters:
      - name: memory_key
        in: path
        required: true
        schema:
          type: string
          title: Memory Key
      - name: wallet_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallet Id
      - name: data_wallet_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Data Wallet Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      - name: include_content
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Include Content
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryContextList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /memory/contexts/{uid}:
    patch:
      tags:
      - memory
      summary: Update Memory Context
      operationId: update_memory_context_memory_contexts__uid__patch
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
          title: Uid
      - name: wallet_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallet Id
      - name: data_wallet_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Data Wallet Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryContextUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryContextItem'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_upload_memory_memory_upload_post:
      properties:
        wallet_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallet Id
        data_wallet_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Data Wallet Id
        app_id:
          type: string
          title: App Id
        session_id:
          type: string
          title: Session Id
        filename:
          anyOf:
          - type: string
          - type: 'null'
          title: Filename
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        summary_threshold:
          anyOf:
          - type: integer
          - type: 'null'
          title: Summary Threshold
        run_push:
          anyOf:
          - type: string
          - type: 'null'
          title: Run Push
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
      - app_id
      - session_id
      - file
      title: Body_upload_memory_memory_upload_post
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MemoryContextItem:
      properties:
        uid:
          type: string
          title: Uid
        memory_key:
          type: string
          title: Memory Key
        wallet_id:
          type: string
          title: Wallet Id
        app_id:
          type: string
          title: App Id
        role:
          type: string
          title: Role
        url:
          type: string
          title: Url
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        content_sha256:
          anyOf:
          - type: string
          - type: 'null'
          title: Content Sha256
        qa_count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Qa Count
        is_summarized:
          anyOf:
          - type: integer
          - type: 'null'
          title: Is Summarized
        created_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Updated At
        content:
          anyOf:
          - type: string
          - type: 'null'
          title: Content
      type: object
      required:
      - uid
      - memory_key
      - wallet_id
      - app_id
      - role
      - url
      title: MemoryContextItem
    MemoryContextList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/MemoryContextItem'
          type: array
          title: Items
        total:
          anyOf:
          - type: integer
          - type: 'null'
          title: Total
      type: object
      title: MemoryContextList
    MemoryContextUpdateRequest:
      properties:
        role:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
          description: user|assistant|system 或自定义
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          description: 记忆条目描述
      type: object
      title: MemoryContextUpdateRequest
    MemoryDeleteResponse:
      properties:
        status:
          type: string
          title: Status
        memory_key:
          type: string
          title: Memory Key
        deleted_contexts:
          type: integer
          title: Deleted Contexts
          default: 0
        deleted_vectors:
          type: integer
          title: Deleted Vectors
          default: 0
        deleted_files:
          type: integer
          title: Deleted Files
          default: 0
        deleted_session:
          type: boolean
          title: Deleted Session
          default: false
      type: object
      required:
      - status
      - memory_key
      title: MemoryDeleteResponse
    MemoryPushRequest:
      properties:
        wallet_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallet Id
          description: 用户钱包 ID（兼容字段；推荐使用 Authorization）
        data_wallet_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Data Wallet Id
          description: 可选：数据归属钱包 ID（业务用户）。默认等于 wallet_id。
        app_id:
          type: string
          title: App Id
          description: 业务插件 ID
        session_id:
          type: string
          title: Session Id
          description: 业务侧会话 ID
        filename:
          type: string
          title: Filename
          description: 业务上传的 session 历史文件名
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          description: 文件说明（可选）
        summary_threshold:
          anyOf:
          - type: integer
          - type: 'null'
          title: Summary Threshold
          description: 摘要触发阈值（可选，覆盖插件配置）
      type: object
      required:
      - app_id
      - session_id
      - filename
      title: MemoryPushRequest
    MemoryPushResponse:
      properties:
        status:
          type: string
          title: Status
        messages_written:
          type: integer
          title: Messages Written
        metas:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Metas
      type: object
      required:
      - status
      - messages_written
      title: MemoryPushResponse
    MemorySessionItem:
      properties:
        memory_key:
          type: string
          title: Memory Key
        wallet_id:
          type: string
          title: Wallet Id
        app_id:
          type: string
          title: App Id
        session_id:
          type: string
          title: Session Id
        created_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Updated At
        message_count:
          type: integer
          title: Message Count
          default: 0
        last_message_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Message At
      type: object
      required:
      - memory_key
      - wallet_id
      - app_id
      - session_id
      title: MemorySessionItem
    MemorySessionList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/MemorySessionItem'
          type: array
          title: Items
        total:
          anyOf:
          - type: integer
          - type: 'null'
          title: Total
      type: object
      title: MemorySessionList
    MemoryUploadResponse:
      properties:
        bucket:
          type: string
          title: Bucket
        key:
          type: string
          title: Key
        source_url:
          type: string
          title: Source Url
        filename:
          type: string
          title: Filename
        size_bytes:
          type: integer
          title: Size Bytes
        content_sha256:
          type: string
          title: Content Sha256
        data_wallet_id:
          type: string
          title: Data Wallet Id
        push_result:
          anyOf:
          - $ref: '#/components/schemas/MemoryPushResponse'
          - type: 'null'
      type: object
      required:
      - bucket
      - key
      - source_url
      - filename
      - size_bytes
      - content_sha256
      - data_wallet_id
      title: MemoryUploadResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...